requirements.txt
# Core scientific computing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Hi-C specific packages
cooler>=0.9.0
cooltools>=0.5.0
coolpuppy>=1.1.0
bioframe>=0.4.0
pairix>=0.3.7

# R integration
rpy2>=3.5.0

# Genomics tools
pybedtools>=0.9.0
pyBigWig>=0.3.18
pysam>=0.19.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.60.0
click>=8.0.0
colorlog>=6.6.0

# Data handling
h5py>=3.6.0
tables>=3.7.0
openpyxl>=3.0.9
xlrd>=2.0.1

# Parallel processing
joblib>=1.1.0
multiprocessing-logging>=0.3.4

# Statistical analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Network analysis
networkx>=2.6.0